pyproject.toml
[project]
name = "distillkit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.4.0",
    "bert-score>=0.3.13",
    "datasets>=3.3.1",
    "deepspeed>=0.16.4",
    "evaluate>=0.4.3",
    "hf-transfer>=0.1.9",
    "ninja>=1.11.1.3",
    "packaging>=24.2",
    "python-dotenv>=1.0.1",
    "rouge-score>=0.1.2",
    "safetensors>=0.5.2",
    "setuptools>=75.8.0",
    "torch>=2.6.0",
    "transformers>=4.49.0",
    "trl>=0.15.1",
    "wheel>=0.45.1",
]
